Implemented PASV command logic. Fixed few utils functions.

+++ lib/netwrk/ip.h
@@ -2,4 +2,4 @@
 
 #include <stddef.h>
 
-#define NK_MAX_IPV4_LEN    ((size_t)16)
+#define NK_MAX_IPV4_LEN    16
